Make support files Python 2 & 3 compatible
This makes the files under support/ mostly Python 3, with backward compatibility for Python 2.

Changes include

- `from __future__ import print_function` and `print(...)`

- imports with `ModuleNotFoundError` or `ImportError` catches(*)

- `.iteritems()` and `.itervalues()` replacead by `items()` and
  `values()` (this assumes there are no large dicts involved).

- `map` replaced by a list comprehension.

- a few specific changes, such as the `func_code` vs. the `__code__` attribute.

(*) Which exception is tested for, depends on what is the easiest to
test for. `ModuleNotFoundError` should be caught for if the first
attempted imports are for Python 2, `ImportError` should be caught for
if the attempted imports are for Python 3 instead.