Support `make *.code` for Python 3
Build parts for 'framework' and a single code are now the same for Python 2 and 3.

For Python 3, the build is out-of-place, in a temporary build
directory. To be able to use the supporting / core packages converted
to Python 3, `setup.py build_py` is run before building a single code.

To be able to fully use a built code, the PYTHONPATH should be set to
the build directory (and include the Python 3 compiled supporting /
core packages directory as well), which would include something like
the `build/lib` and `build/temp...` directories.

In the future, *installing* a single code should be the default, which
avoids adjusting the PYTHONPATH: just building the code only works
in-place, which is on its way out of being the default.