Support `make *.code` for Python 3
Build parts for 'framework' and a single code are now the same for Python 2 and 3.

This does rely more on a hack than proper Python 3 support: because
the Python 2 Amuse code code needs to be compiled to Python 3, but not
in place (to not mess up the (git) source), the necessary Python files
are copied to the temporary build directory, and then compiled. This
means that, to actually use the code, one should set their PYTHONPATH
to the relevant build directory (this happens automatically when
building though, i.e., when using `make <something>.code`).

Compared to building all codes at once
(support/setup_codes.py/BuildCodes or BuildLibraries), which uses
copy_codes_to_build_dir, this copies once, but avoids copying again if
a file is already found in the build dir (and presumed compiled for
the correct Python version).

This work-around can be avoided in the future in two ways:

- the core Amuse code is Python 3, and backwards compatible with
  Python 2. In that case, no copying of files is necessary.

- the code Amuse code is its own package, and becomes an essential
  requirement to install before using the generic Amuse package. In
  this case, the installation of the core Amuse code will have ensure
  the code is already compiled for the relevant Python version.