Implement opportunistic deadlock avoidance for worker crashes.
Amuse has a tendency to deadlock when the remote worker crashes, due to the
worker(s) never calling MPI_Finalize or MPI_Abort. This change installs (only
on POSIX platforms and only on thread-safe MPI initialisations) a signal
handler that traps some of the more common "crash" signals (SIGABRT, SIGBUS,
SIGILL, SIGINT, SIGQUIT, SIGSEGV, and SIGTERM) and attempts to call MPI_Abort()
before dying.

This is not sufficient to avoid deadlock, as the MPI_Abort() never notifies
the MPI_Comm_spawn parent that the run is aborted. The parent (i.e., amuse) is
generally blocked in an MPI_Wait, waiting for a return message from a function
call, due to the way mpi4py is implemented amuse becomes (nearly)
uninterruptible by keyboard interrupts (SIGINT), nor does it detect the
MPI_Abort in the worker.

To work around this, the worker sends a non-blocking all 0 message to the
parent (waking it up from the MPI_Wait), this message fails the call and
function id checks on lines 1149 and 1152 of amuse/rfi/channel.py, resulting in
a call to MPI_Comm_disconnect in the parent. The worker also calls
MPI_Comm_disonnect and only then calls MPI_Abort to end the other worker tasks.

This way both the workers and clients abort when an error/crash happens. This
doesn't always work (the disconnect could block, some other MPI error could
happen while trying to send/abort, etc.), but since the current behaviour is to
just deadlock, we're not off any worse than before and when it does work things
are much more pleasant to test with.