Tweak mesa2208 inlist loading
The AMUSE_inlist is tweaked so that it looks for a local file called
"inlist" (we also patch out the error so if this local inlist does not
exist its not an error). This makes it easier to provide the additional
input parameters that amuse does not expose in its interface available
to users, without needing to edit the main AMUSE_inlist file.