Revert "Accept system CUDA in a conda env, it's often better"
This reverts commit 743c0d5fc7c769d77c3a7c58f5e5a589e0483345. It turns
out that a system (e.g. apt-installed) CUDA doesn't work with the Conda
compilers. This does work on a cluster with modules though, and it
avoids potential compatibility issues between the CUDA we compiled with
and the (potentially older) CUDA driver on the system, so we'll have to
figure out a different solution for that.