Fixes #1195
Apparently, this line of code:
"
  float inv_ds  = rsqrt(ds2 + EPS2) * (__float_as_int(pos_i.w.y) != __float_as_int(pos_j.w.y));
"
which should result in "inv_ds==0" for "i==j" results in "inv_ds==NaN" for "i==j" for modern CUDA versions.
We suspect that for older CUDA versions this worked well and resulted in "inv_ds==0" for "i==j", i.e. for fictious self-gravitation - which we exclude, of course.
Most likely, older CUDA versions did not comply with IEEE 754.